requirements.txt
Flask==2.3.2
Flask_JWT_Extended==4.5.2
Flask_RESTful==0.3.10
flask_sqlalchemy==3.0.5
flask_swagger==0.2.14
flask_swagger_ui==4.11.1
python-dotenv==1.0.0
